update gps code

+++ Project1/3년차/gps/gps_with_gyro2.cpp
@@ -306,9 +306,9 @@ float get_magnetic(char *chrBuf, int num)
     }
 }
 //0x56 Atmospheric data 추출
-float get_atmospheric(char *chrBuf, int num)
+double get_atmospheric(char *chrBuf, int num)
 {
-    float press; float h;
+    double press; double h;
     signed int tmp[8];// int로 변경
     unsigned char i;
 
@@ -317,11 +317,11 @@ float get_atmospheric(char *chrBuf, int num)
     }
 
     if (num==1){
-        press = (float)((tmp[3]<<24)|(tmp[2]<<16)|(tmp[1]<<8)|tmp[0]);
+        press = (double)((tmp[3]<<24)|(tmp[2]<<16)|(tmp[1]<<8)|tmp[0]);
         return press;
     }
     else if (num==2){
-        h = ((float)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]))/100;
+        h = ((double)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]))/100;
         return h;
     }
     else{
@@ -329,9 +329,9 @@ float get_atmospheric(char *chrBuf, int num)
     }
 }
 //0x57 GPS data 추출
-float get_gpsData(char *chrBuf, int num)
+double get_gpsData(char *chrBuf, int num)
 {
-    float lon; float lat;
+    double lon; double lat;
     signed int tmp[8];// int로 변경
     unsigned char i;
 
@@ -340,11 +340,11 @@ float get_gpsData(char *chrBuf, int num)
     }
 
     if (num==1){
-        lon = (float)((tmp[3]<<24)|(tmp[2]<<16)|(tmp[1]<<8)|tmp[0]);
+        lon = (double)((tmp[3]<<24)|(tmp[2]<<16)|(tmp[1]<<8)|tmp[0]);
         return lon;
     }
     else if (num==2){
-        lat = ((float)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]));
+        lat = ((double)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]));
         return lat;
     }
     else{
@@ -352,9 +352,9 @@ float get_gpsData(char *chrBuf, int num)
     }
 }
 //0x58 Ground Speed data 추출
-float get_groundSpeed(char *chrBuf, int num)
+double get_groundSpeed(char *chrBuf, int num)
 {
-    float gh; float gy; float gv;
+    double gh; double gy; double gv;
     signed short tmp[8];
     unsigned char i;
 
@@ -363,15 +363,15 @@ float get_groundSpeed(char *chrBuf, int num)
     }
 
     if (num==1){
-        gh = ((float)((tmp[1]<<8)|tmp[0]))/10;
+        gh = ((double)((tmp[1]<<8)|tmp[0]))/10;
         return gh;
     }
     else if (num==2){
-        gy = ((float)((tmp[3]<<8)|tmp[2]))/10;
+        gy = ((double)((tmp[3]<<8)|tmp[2]))/10;
         return gy;
     }
     else if (num==3){
-        gv = ((float)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]))/1000;
+        gv = ((double)((tmp[7]<<24)|(tmp[6]<<16)|(tmp[5]<<8)|tmp[4]))/1000;
         return gv;
     }
     else{
@@ -448,9 +448,9 @@ float ax; float ay; float az; float t; //0x51
 float wx; float wy; float wz; //0x52
 float roll; float pitch; float yaw;//0x53
 float mx; float my; float mz;//0x54
-float press; float h; //0x56
-float lon; float lat; float lon_dd; float lat_dd; float lon_mm; float lat_mm;//0x57
-float gh; float gy; float gv;//0x58
+double press; double h; //0x56
+double lon; double lat; double lon_dd; double lat_dd; double lon_mm; double lat_mm;//0x57
+double gh; double gy; double gv;//0x58
 float q0; float q1; float q2; float q3;//0x59
 float sn; float pdop; float hdop; float vdop;//0x5a
